+++ app/Http/Controllers/Admin/ProductController.php
@@ -9,6 +9,7 @@ use App\Models\Image;
 use App\Http\Requests\ProductRequest;
 use App\Http\Requests\EditProductRequest;
 use DB;
+use Auth;
 
 class ProductController extends Controller
 {
@@ -20,34 +21,38 @@ class ProductController extends Controller
      */
     public function store(ProductRequest $request)
     {
-        $products = Product::where('product_information_id', $request->product_information_id)->get();
-        foreach ($products as $product) {
-            if ($product->color_id == $request->color_id) {
-                return redirect()->back()->with('error_color', 'Thất bại! Loại sản phẩm đã tồn tại');
+        if (Auth::guard('admin')->user()->can('create', Product::class)) {
+            $products = Product::where('product_information_id', $request->product_information_id)->get();
+            foreach ($products as $product) {
+                if ($product->color_id == $request->color_id) {
+                    return redirect()->back()->with('error_color', 'Thất bại! Loại sản phẩm đã tồn tại');
+                }
             }
-        }
-        DB::beginTransaction();
-        try {
-            $product = Product::create([
-                'product_information_id' => $request->product_information_id,
-                'color_id' => $request->color_id,
-                'quantity' => 0,
-                'unit_price' => $request->unit_price,
-            ]);
-            foreach ($request->file('images') as $file) {
-                Image::create([
-                    'product_id' => $product->id,
-                    'image_link' => 'storage/' . $file->getClientOriginalName(),
+            DB::beginTransaction();
+            try {
+                $product = Product::create([
+                    'product_information_id' => $request->product_information_id,
+                    'color_id' => $request->color_id,
+                    'quantity' => 0,
+                    'unit_price' => $request->unit_price,
                 ]);
-                $file->move('storage', $file->getClientOriginalName());
+                foreach ($request->file('images') as $file) {
+                    Image::create([
+                        'product_id' => $product->id,
+                        'image_link' => 'storage/' . $file->getClientOriginalName(),
+                    ]);
+                    $file->move('storage', $file->getClientOriginalName());
+                }
+                DB::commit();
+
+                return redirect()->back()->with('done', 'Thao tác thành công!');
+            } catch (Exception $e) {
+                DB::rollBack();
+
+                return $e->getMessage();
             }
-            DB::commit();
-
-            return redirect()->back()->with('done', 'Thao tác thành công!');
-        } catch (Exception $e) {
-            DB::rollBack();
-
-            return $e->getMessage();
+        } else {
+            abort(401);
         }
     }
 
@@ -60,12 +65,16 @@ class ProductController extends Controller
      */
     public function update(EditProductRequest $request, $id)
     {
-        $product = Product::findOrFail($id);
-        $product->update([
-            'unit_price' => $request->unit_price,
-        ]);
+        if (Auth::guard('admin')->user()->can('update', Product::class)) {
+            $product = Product::findOrFail($id);
+            $product->update([
+                'unit_price' => $request->unit_price,
+            ]);
 
-        return redirect()->back()->with('done', 'Thao tác thành công!');
+            return redirect()->back()->with('done', 'Thao tác thành công!');
+        } else {
+            abort(401);
+        }
     }
 
     /**
@@ -76,20 +85,24 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        DB::beginTransaction();
-        try {
-            $product = Product::findOrFail($id);
-            $product->delete();
-            foreach ($product->images as $image) {
-                $image->delete();
+        if (Auth::guard('admin')->user()->can('delete', Product::class)) {
+            DB::beginTransaction();
+            try {
+                $product = Product::findOrFail($id);
+                $product->delete();
+                foreach ($product->images as $image) {
+                    $image->delete();
+                }
+                DB::commit();
+
+                return redirect()->back()->with('done', 'Thao tác thành công!');
+            } catch (Exception $e) {
+                DB::rollBack();
+
+                return $e->getMessage();
             }
-            DB::commit();
-
-            return redirect()->back()->with('done', 'Thao tác thành công!');
-        } catch (Exception $e) {
-            DB::rollBack();
-
-            return $e->getMessage();
+        } else {
+            abort(401);
         }
     }
 }
